# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchDeleteRequest": {
        "description": "Soft-delete multiple media records.",
        "properties": {
          "media_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 500,
            "title": "Media Ids",
            "type": "array"
          }
        },
        "required": [
          "media_ids"
        ],
        "title": "BatchDeleteRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RestoreMediaRequest": {
        "description": "Undo a media soft-delete.",
        "properties": {
          "deleted_at": {
            "title": "Deleted At",
            "type": "string"
          },
          "media_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 500,
            "title": "Media Ids",
            "type": "array"
          }
        },
        "required": [
          "media_ids",
          "deleted_at"
        ],
        "title": "RestoreMediaRequest",
        "type": "object"
      },
...
      "RunSelectedRequest": {
        "description": "Queue a pipeline run for specific media IDs.",
        "properties": {
          "confidence_threshold": {
            "default": 0.2,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Threshold",
            "type": "number"
          },
          "media_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Media Ids",
            "type": "array"
          },
          "steps": {
            "default": [
              "speciesnet"
            ],
            "description": "Pipeline step types to run",
            "items": {
              "type": "string"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "media_ids"
        ],
        "title": "RunSelectedRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "An organisation API key, ww_live_...",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "V2 backend \u2014 async job system, model conversion",
    "title": "Wildlife Watcher API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/media/batch": {
      "delete": {
        "description": "Soft-delete multiple media records (sets deleted_at).\n\nRuns as the requesting user (not the service role), so RLS decides which rows change:\na photo's uploader while they hold at least project_member on its project\n(ww-backend #277), or a project_admin of the project. Any other id, including a plain\nmember's request for a photo someone else uploaded, changes 0 rows without an error, as\ndoes an id already deleted. ``deleted_ids`` lists the rows that changed and\n``skipped_ids`` the rest; ``deleted`` counts ``deleted_ids``.",
        "operationId": "batch_delete_media_api_media_batch_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Delete Media",
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/batch/restore": {
      "post": {
        "description": "Undo a media delete, clearing ``deleted_at`` where it equals the given timestamp. Runs as the\nuser, so RLS applies the delete rule: the photo's uploader while they hold at least\nproject_member, or a project_admin of the project; other ids change 0 rows without an error.\nScoping by the exact timestamp restores only the photos removed in that delete.\n``restored_ids`` lists the rows that changed and ``skipped_ids`` the rest.",
        "operationId": "batch_restore_media_api_media_batch_restore_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreMediaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Restore Media",
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/registry/{deployment_id}": {
      "get": {
        "description": "Paginated media list with pre-resolved URLs, null where only the private original exists.",
        "operationId": "media_registry_api_media_registry__deployment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Media Registry",
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/run-selected": {
      "post": {
        "description": "Queue a pipeline run for specific media IDs (not the full deployment).\n\nLooks up the deployment_id from the first media record, then runs the\npipeline with a media_ids filter. The lookup runs as the requesting user\nso RLS limits it to media in their own projects.",
        "operationId": "run_pipeline_selected_api_media_run_selected_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSelectedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Pipeline Selected",
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/thumbnails/{deployment_id}": {
      "post": {
        "description": "Enqueue an async job to generate thumbnails/previews for existing media.",
        "operationId": "enqueue_thumbnail_backfill_api_media_thumbnails__deployment_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enqueue Thumbnail Backfill",
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/{media_id}/image": {
      "get": {
        "description": "Serve a media image by resolving its file_path through the provider registry.\n\n- ``size=thumb``: fast low-res preview for the thumbnail grid\n- ``size=full``: high-res download for the detail panel (on-demand)\n\nUses the authenticated user's Supabase client so RLS enforces\nproject membership \u2014 users can only view media they have access to.",
        "operationId": "get_media_image_api_media__media_id__image_get",
        "parameters": [
          {
            "in": "path",
            "name": "media_id",
            "required": true,
            "schema": {
              "title": "Media Id",
              "type": "string"
            }
          },
          {
            "description": "Image size: thumb (grid) or full (detail)",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": "thumb",
              "description": "Image size: thumb (grid) or full (detail)",
              "enum": [
                "thumb",
                "full"
              ],
              "title": "Size",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Media Image",
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/{media_id}/resolve": {
      "get": {
        "description": "Return a URL a plain ``<img>`` can load for a media item, regardless of storage provider.\n\n``url`` is null when the photo has no rendition yet and its original is private.",
        "operationId": "resolve_media_url_api_media__media_id__resolve_get",
        "parameters": [
          {
            "in": "path",
            "name": "media_id",
            "required": true,
            "schema": {
              "title": "Media Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": "thumbnail",
              "enum": [
                "thumbnail",
                "preview",
                "original"
              ],
              "title": "Size",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Media Url",
        "tags": [
          "media"
        ]
      }
    },
...
  }
}